Refactor configuration
Why:

* Following on #362, we're moving all files to use the file naming
  convention of having each file named after its default export, which
  means camelCase for functions;
* Taking advantage of the fact we're moving the files, we're also
  nesting files related to `gatsby-config` to its own subdirectory;
* `gatsby-ssr` config is just a placeholder which can be removed;
* Since the configuration files also rely on some common logic for
  generating the URLs, and these are too specific to be extractable to
  `@subvisual/utils`, we're moving them to `lib/` since these are not
  exclusive to the site's source;
* While refactoring this configuration it also became obvious the SEO
  component is assuming a site-wide default for the `lang` attribute,
  which can be set in the site metadata instead.
