+++ config/gatsby/siteMetadata.js
@@ -1,0 +1,12 @@
+const buildURL = require("../../lib/buildURL")
+
+module.exports = {
+  description: "We nurture ideas that empower people",
+  image: "/images/meta-image.jpg",
+  lang: "en",
+  title: "Subvisual",
+  twitter: {
+    creator: "@subvisual",
+  },
+  url: buildURL().toString(),
+}
