--- config/site_metadata.js
@@ -1,7 +1,0 @@
-module.exports = {
-  title: "Subvisual",
-  description: "We nurture ideas that empower people",
-  url: "https://subvisual.com",
-  image: "/images/meta-image.jpg",
-  twitterUsername: "@subvisual",
-}
