+++ gatsby-node.js
@@ -1,4 +1,5 @@
-exports.createPages = require("./config/node/create_pages")
-exports.createSchemaCustomization = require("./config/node/create_schema_customization")
-exports.onCreateNode = require("./config/node/on_create_node")
-exports.onCreateWebpackConfig = require("./config/node/on_create_webpack_config")
+exports.createPages = require("./config/node/createPages")
+exports.createSchemaCustomization = require("./config/node/createSchemaCustomization")
+exports.onCreateNode = require("./config/node/onCreateNode")
+exports.onCreateWebpackConfig = require("./config/node/onCreateWebpackConfig")
+exports.onPreInit = require("./config/node/onPreInit")
